Improve Testing (#142)
* feat: Add new Page for testing

Add a new page with a Scaffold and add a simple test on this composable

* test: Add test for content inside the button

* test: Improve SwipePage test and add Row tests

Add useUnmergedTree to true to only initialize the SwipePage composable at beginning

* test: Use assertExists instead of assertDisplayed

Use assertExists instead of assertDisplayed and modify mock name for more readability

* test: Use assertExists instead of assertDisplayed

Use assertExists instead of assertDisplayed and modify mock name for more readability

* fix: fluidity bug when dragging

Improve fluidity when dragging by updating the image only at the end of the animation and improve tests by checking if the image change after swiping

* fix: Correct waitForIdle issue when running cis

Use runBlocking with awaitIdle instead of only waitForIdle to try fixing the bug that waitForIdle doesn't wait long enough

* fix: Correct waitForIdle issue by waiting current Recipe changes

Wait for current recipe is defined and not null before checking if the nextRecipe function have been called

* fix: Force onSuccess to run on principal thread

Fix issue of ci not run asynchronous event the same as local by using composeTestRule.runOnUiThread {} to ensure callbacks are executed on the main thread

* fix: Resolve issue while currentRecipe doesn't update when testing with ci
Use thenAnswer with a doAnswer to force executing the onSuccess directly when random is called

* fix: Resolve issue of coroutine testing

use runTest to test coroutines. Use advanceUntilIdle() to let all coroutine finishing before continuing testing

* fix: Resolve issue of coroutine testing with runCurrent

use runCurrent to test coroutines. Test if fetchRandomRecipes executes correctly

* fix: Resolve issue of coroutine testing with job and collectLatest

use job and wait for loading set to false until testing to test coroutine

* fix: Call nextRecipe() outside of coroutine

* test: Add test to if nextRecipe() correctly called

Add tests with only nextRecipe(), with and without job/advanceUntilIdle. Testing why worked on local and not on ci. Update SwipePage.kt to call nextRecipe correctly

* test: Add manually clock control
Add manually clock control to control when dragging event is finished before checking if the current recipe have been updated

* test: Increase swipe size to ensure action is triggered

Adjust the swipe size to guarantee it is large enough to consistently trigger the intended action.

* test: Increase swipe size to ensure action is triggered

Adjust the swipe size to guarantee it is large enough to consistently trigger the intended action.

* feat: Load second recipe in background to enhance UI experience

- Implement a second recipe to improve UI responsiveness
- Add test to confirm second recipe is displayed as expected

* feat: add Recipe overview access, portrait mode, and back navigation

- Enabled access to Recipe Overview when tapping on description
- Tested new feature functionality
- Restricted screen orientation to portrait in manifest
- Added back navigation to RecipeOverview screen
- Removed unused TestingScreen and associated test

* doc: Clarify code documentation for better understanding

* chore: delete unnecessary commented line

---------

Co-authored-by: Adrien Bader <Adrien Bader>