Feature/image storage (#155)
* feat: add Image data class

Added the Image data class which will used for the storage of the images in Firebase Storage. In addition, in the same file, there is an enum class which will serve to identify what each image describes.

* feat: add ImageRepository interface

Added the ImageRepository interface with the functions which will be used to interact with the Firestore Storage

* feat: add implementation of firebase storage interaction methods

Added the implementation of the methods that will interact with the Firebase Storage. In addition, the data class of the Image was deleted and some parameters of the methods in the interface were changed.

* feat: add implementation of firebase storage interaction methods

Added the implementation of the methods that will interact with the Firebase Storage. In addition, the enum class ImageType was moved in ImageRepositoryFirebase.kt file following the deletion of the file Image.kt, and some parameters of the methods in the interface were changed.

* fix: modify the parameters of methods

Modified the parameters of the implemented methods in order to allow naming of the different files in the directory of a certain recipe for example. In addition, a constant was renamed and another one was removed, and the documentation of the methods was changed.

* refactor: format with ktfmt

Formated the core using the ktfmtFormat from ktfmt.