Upgrade login API version (#165)
* fix: Upgrade login API version
Upgrade the version of the login API to fix login issues. To be tested

* test: update SignIn tests to match with the upgraded login

* feat(ui): Update UI components for improved user experience

Add a cook image and multiples recipes to match the Figma's design. The recipes are animated but the cook is fixed.

* feature: Upgrade login with the bottom login UI of Google

Use bottom login of google with GetGoogleIdOption, add cook, salad and taco's image

* feat: Add loading animation after sign in

- Implemented LoadingAnimation composable to display falling fruits.
- Added FallingFruit class to handle individual fruit properties and animation.
- Created tests to ensure proper rendering and functionality of the loading animation.
- Call LoadingAnimation composable in SignInScreen upon successful sign-in to enhance user experience.

* doc: Add constants into C.kt and add documentation
- Moved all constant values from various files into C.kt for better organization and maintainability.
- Updated code comments to provide clear documentation.
- Ensured that all references to constants in the project are updated accordingly.

* fix: Add time delay constant to fix fruit fall

Fix fruit falling to fast by adding a specific TIME_DELTA constant to have the correct DELAY to simulate fps

* test: Add Sign-In test for stopping SignIn

This commit introduces a test to verify that the Sign-In process handles invalid credentials correctly.

* refactor: Rename SignInTest to SignInScreenTest for consistency and simplify SignInScreen

This commit renames SignInTest to SignInScreenTest to improve consistency across the test suite. Additionally, it removes the handleSignIn method, consolidating the sign-in logic into a single method to enhance readability and maintainability.

* chore: Compress animation images to reduce file size

- Compressed images used in animations to improve load times and performance.
- No visual quality changes; only file size reduced.

---------

Co-authored-by: Adrien Bader <Adrien Bader>