feature/link-account-recipes (#195)
* fix: show liked recipes on account page

Fixed a bug where the liked recipes were not showing up on the Account screen because of the viewmodel

* feature: display selected recipe on account screen

Added the possibility to select a recipe on the account page and view it. In addition, the tests were modified to accompany the new modifications.

* feature: remove liked recipes

Added the option to let the user to removed a certain liked recipe.

* test: modify existing tests

Modified existing test files to comply with the changes made

* test: add account screen tests

Added account screen tests in order to test the newly added features

* test: add additional userViewModel tests

Added additional userViewModel tests that test the new code added in the class

* test: add forgotten userViewModel test

Added forgotten userViewModel test that tests the log of the error when an ingredient is not found in the database.

* refactor: add constants to C file

Added some the test tags as well as the different constants to the C file.

* refactor: add constants to C file and string file

Added some constants in the C.kt and strings.xml that were found in the code. Same goes for some test tags in the tests file as well as some dimensions of items

* test: add tests for new feature in the account screen

Added some tests for the pop up message that comes up when the user tries to unlike a recipe on the account page

* feat: add pop up when unliking a recipe on the account page

Added a pop up menu that comes up when the user tries to unlike a recipe and that confirms the fact that he want to perform the action.

* test: add test for the profile picture

Added a test to see if a profile picture is displayed if the profile picture url is not empty or null

* fix: add correct message for pop up

The pop up message was changed to the correct one because it had the wrong one.

* feature: update user in database

Added in each update of a user's description (e.g., username) a call to updateCurrentUser that updates the user in the database

* refactor: add constants to C.kt

Added some found constants to the constants file.

* refactor: refactor with ktfmt

Code was refactored using ktfmt