Tests/end to end (#211)
* add feature search screen with search bar according to Figma design

* add feature search screen with search bar according to Figma design

* add feature search screen with search bar according to Figma design

* add feature search screen with search bar according to Figma design

* ktfmt for CI

* fix duplicate filter icon file

* fix end to end tests

* fix end to end test

* fix CI issue

* pass display tests

* delete filter functions because not needed anymore

* add changes to searchBar composable

* fix: show liked recipes on account page

Fixed a bug where the liked recipes were not showing up on the Account screen because of the viewmodel

* add changes to searchScreen composable and testing the display

Use of utils composable functions
Using testTag values for testing the display
Centering the searchBar

* feature: display selected recipe on account screen

Added the possibility to select a recipe on the account page and view it. In addition, the tests were modified to accompany the new modifications.

* feature: remove liked recipes

Added the option to let the user to removed a certain liked recipe.

* test: modify existing tests

Modified existing test files to comply with the changes made

* test: add account screen tests

Added account screen tests in order to test the newly added features

* test: add additional userViewModel tests

Added additional userViewModel tests that test the new code added in the class

* test: add forgotten userViewModel test

Added forgotten userViewModel test that tests the log of the error when an ingredient is not found in the database.

* refactor: add constants to C file

Added some the test tags as well as the different constants to the C file.

* implement naviagtion test when clicking on a recipeCard in SearchScreen

* implement naviagtion test when clicking on a recipeCard in SearchScreen

* add filer to searchScreen

* making searchBar single line search bar

* fix package name error

* fix package name error in MainActivity

* apply ktfmt

* changing some testTag for precision

* fix some tests

* add search screen tests

* add some changes to the end to end tests

* remove: changes from searchscreen branch

All of these changes are specific to searchscreen and should be removed.

* refactor: add constants to C file and string file

Added some constants in the C.kt and strings.xml that were found in the code. Same goes for some test tags in the tests file as well as some dimensions of items

* test: add tests for new feature in the account screen

Added some tests for the pop up message that comes up when the user tries to unlike a recipe on the account page

* feat: add pop up when unliking a recipe on the account page

Added a pop up menu that comes up when the user tries to unlike a recipe and that confirms the fact that he want to perform the action.

* test: add test for the profile picture

Added a test to see if a profile picture is displayed if the profile picture url is not empty or null

* fix: add correct message for pop up

The pop up message was changed to the correct one because it had the wrong one.

* add: missing testtags

* add: test create recipe end to end test

* feature: update user in database

Added in each update of a user's description (e.g., username) a call to updateCurrentUser that updates the user in the database

* add filter test and fix likedrecipes test

* fix likedRecipe test

* delete unused imports

* ktfmt formatting

* fix: incorrect line in fake navhost

For some reason, there was a line not representative of the main activity nav graph

* Delete app/_gitignore

* Delete app/src/main/res/drawable/search.xml

* Delete app/src/main/res/drawable/back.xml

* Delete app/src/main/res/drawable/plateswipelogo.xml

---------

Co-authored-by: caprasnitzel <ricy_alex@yahoo.com>
Co-authored-by: RagdollGoDown <zenooby2002@gmail.com>
Co-authored-by: André Cadet <86296692+RagdollGoDown@users.noreply.github.com>