fix: UserViewModelTest incorrectly checking emptylists
The UserViewModelTest would not use hardcoded values, but would check if a list String was equal to a list Recipe to check if it was empty.
Also it would really on the lists being empty for testing, so now it uses the sample class user and can be adapted to any sample user.