Feature/enhance aggregator search (#240)
* change: made it impossible to add ingredients with the same barcode in FirestoreIngredientRepository
Now when an element exists in the firestore db with the same barcode it will either update it, if the updateIfExists param is true, or add it

There is also a new excepetion for when the ingredient has no barcode and someone is attempting to add it to the db.

* change: AggregatorIngredientRepository will not add ingredients into firestore db

This is due to the unfortunate difficulty and unreliability when adding new ingredients to the firestore db. It isn't worth our time to go through all the bugs as any use of firestore is in the end done in the get function once our application actually uses the ingredient.

* add: isSearching stateflow to IngredientViewModel

This is true when the viewmodel is fetching asynchronously a list of ingredients. It simplifies the process in the ingredient search screen as it now doesn't rely on the launchedEffect method. This method wouldn't trigger on certain changes and leave the screen perpetually loading.

* Revert "change: made it impossible to add ingredients with the same barcode in FirestoreIngredientRepository"

This reverts commit 33cb0a7f01243a2ea5b03fb2c4bf2d3aae02bccb.

* remove: debugging log in search and unused constant

* remove: compilation errors due to old test version of add in FirestoreIngredientRepo

* change: ktfmt AggregatorIngredientRepository

* fix: missing parameter in mock function