Feature/fridge backend (#237)
* feat: add,remove, update ingredients in fridge

Modifies methods that interact with the user viewmodel and the database in order to fetch and update the list on ingredients that the user has, and added an update method that updates an ingredient with the given quantity.

* test: modify test for compatibility with the modifications

Modified the existing methods to be compatible with the different modifications.

* feat: edit snapshot conversion

Edited the way that the snapshot conversion is done when fetching a user from the database. In addition, the documentation of some methods was added or updated.

* test: add tests for the new modifications

Added and edited tests in order to be compatible with the new modifications. In addition, added some general objects to the SampleClasses.kt

* test: enhance update ingredient

Enhanced the update ingredient method in order to be able to remove an ingredient from the fridge when it reaches a 0 quantity, but also add to the quantity when its the same ingredient.

* test: add tests for new implementations

Added tests which test the new implementations.

* feat: modify implementation based on feedback from review

Made modifications based on the feedback received from the review

* test: modify and add tests based on feedback from review

Modified currents tests and added new ones based on the feedback received from the review