feature/edit account ui (#258)
* feature: link edit account screen to app

Linked the Edit Account Screen to the rest of the app. Meaning that we can go to this screen from the account page and go back.

* feature: change username and profile picture, see email address linked with account

Added the possibility to change the username and the profile picture. In addition, A field was added to see which email address is used for the account.

* feature: change the date of birth of user

Added to the user the possibility to change his birth of date. By default, there is no date

* refactor: ktfmt format

Factored the code using the ktfmt format

* refactor: modularize code

Some parts of the code were modularized in order to obtain better visibility and maintainability

* test: add tests for ner feature

Added test for the newly implemented feature.

* refactor: ktfmt format

Code was formated using ktfmt format

* refactor: add constants to constants file

Added the constants from the implementation to the constants file.

* feat: change date picker pop up

Changed the date picker pop up in order to have a better UX because it introduces a cancel and confirm button which is more intuitive to the user.

* test: add tests for new implementation

Added some tests for the new date picker pop up in order to test it.

* test: add tests for the date picker and new image

Added some tests for choosing a date in the date picker and using an existing profile picture.

* test: add tests for fetch failure case of an image

Added a test in which I test the case were the image fetching fails.

* test: correct after merge conflict

Modified setup of the test after merge conflict in order to work.

* refactor: changes from pull request

Made changes to the code and tests to comply with the feedback from the pull request

* refactor: changes comment about LaunchedEffect

Modified the comment about the LaunchedEffect because it was not clear what was trying to be transmitted

* test: correct tests after merge conflict

Corrected an existing test in order to comply with the modifications found on the main branch