Feature/filter page backend (#257)
* feat: implement filterSearch method in FirestoreRecipesRepository

Implement the filterSearch method. This method takes from the database
nbOfElements * FILTER_RANDOM_FACTOR elements that correspond to the Filter
parameters, shuffles them and applies onSuccess to the first nbOfElements of
this list which adds some randomness.

* feat: add the difficulty filter to the filterSearch function

* feat: add precondition to filterSearch

Limit the limit parameter to positive numbers

* opt: optimize filterSearch function by not recreating a query

* test: add a test for error throwing and normal execution

Add a test for limit <= 0 and a normal execution
that should be improved later

* doc: update documentation of filterSearch function

* doc: update documentation of filterSearch function and comments

* test: add tests for every branch in the filterSearch function

* chore: make listCategories not mandatory

Delete the listCategories from the RecipeRepository
interface due to previous discussions where this
method was moved in the Recipe data class.

* fix: add a forgotten wait for Asynchronous operations in a test

* feat: delete UI related to price

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: delete UI related to price (forgotten)

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: delete UI related to price (forgotten)

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: delete UI related to price (forgotten)

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: delete UI related to price (forgotten)

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: delete UI related to price (forgotten)

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: delete UI related to price (forgotten)

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: delete UI related to price (forgotten)

Since we sayed we will not use the price
This commit deletes all elements of the
Filter where the price is taken.

* feat: adapt testing for new function

Replace searchCategory by filterSearch in the RecipesViewModelTest

* feat: adapt the interfaces RecipeRepository interface to make filterSearch mandatory

* feat: time formatting

* feat: change searchCategory by filterSearch

* feat: change searchCategory by filterSearch

Adapt tests to mock the filterSearch call.

* feat: add function that checks if filter is used

Add a function that checks if some chips are displayed and then
resets the filter or not

* fix: adapt filter to manage better null values

Adapt the filter such a way that null values can exist,
because on our DB, some values are null. Searching will be
better if the values in the filter can be null

* fix: change time range value when change in Slider

The values will be set only if the default values are changed
otherwise they will still be default values.

* feat: make reset button navigate to SwipeScreen

* doc: update documentation

* fix: fix the fetching problem

Fix the problem where time filter could not fetch any recipe by applying
a toInt().toString() to the time max and min.

* fix: take care to reset the filter at init and reset

* fix: fix the display of the time chip

Time chip was displayed every time the filter was applied
even if time was not modified.

* fix: fix the display of the time chip

Time chip was displayed every time the filter was applied
even if time was not modified.

* fix: fix fetching by giving an orderBy

* fix: suppress tests that do not make sense now

* fix: adapt tests to new searchFilter logic

* fix: fix one test

* chore: ktfmt format

* chore: add the missing method to the interface

* chore: suppress listCategories

* fix: code duplication in Interface so made one extend the other

* chore: move return satement out of if

* chore: delete unused parameter

* fix: fix crash when selecting filter

* chore: replace value by constant MINUTES_PER_HOUR

* doc: add documentation to the isLimited function